Move connection things to own namespace.
To keep the connection and processing nicely separated.
