+++ Source/EventHorizon/Consumer/Connections/EventHorizonConnectionFactory.cs
@@ -2,12 +2,11 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using Dolittle.Runtime.EventHorizon.Consumer.Processing;
 using Dolittle.Runtime.Microservices;
 using Dolittle.Runtime.Services.Clients;
 using Microsoft.Extensions.Logging;
 
-namespace Dolittle.Runtime.EventHorizon.Consumer
+namespace Dolittle.Runtime.EventHorizon.Consumer.Connections
 {
     /// <summary>
     /// Represents an implementation of <see cref="IEventHorizonConnectionFactory"/>.
