+++ Source/EventHorizon/Consumer/Connections/IEventHorizonConnectionFactory.cs
@@ -1,10 +1,9 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using Dolittle.Runtime.EventHorizon.Consumer.Processing;
 using Dolittle.Runtime.Microservices;
 
-namespace Dolittle.Runtime.EventHorizon.Consumer
+namespace Dolittle.Runtime.EventHorizon.Consumer.Connections
 {
     /// <summary>
     /// Defines a system that can create instances of <see cref="IEventHorizonConnection"/>.
